refactor the bff structure
