--- bff/src/graphql/status.module.ts
@@ -1,9 +1,0 @@
-import { StatusService } from '../rest/status/status.service'
-import { StatusResolver } from './status/status.resolver'
-import { Module } from '@nestjs/common'
-
-@Module({
-  controllers: [],
-  providers: [StatusResolver, StatusService],
-})
-export class StatusModule {}
