+++ bff/src/modules/status/status.module.ts
@@ -1,0 +1,10 @@
+import { StatusResolver } from './graphql/status.resolver'
+import { StatusController } from './openapi/status.controller'
+import { StatusService } from './status.service'
+import { Module } from '@nestjs/common'
+
+@Module({
+  controllers: [StatusController],
+  providers: [StatusResolver, StatusService],
+})
+export class StatusModule {}
