--- bff/src/rest/status/status.module.ts
@@ -1,9 +1,0 @@
-import { StatusController } from './status.controller'
-import { StatusService } from './status.service'
-import { Module } from '@nestjs/common'
-
-@Module({
-  controllers: [StatusController],
-  providers: [StatusService],
-})
-export class StatusModule {}
